sua lai kpi

+++ app/Http/Controllers/Api/TaskController.php
@@ -107,8 +107,10 @@ class TaskController extends Controller
                 if ($task->stage_id != $stage->id) {
 
                     if ($stage->index > $task->stage->index) {
-                        $a =    Kpi::query()->where('task_id', $task->id)->where('stage_id', $stage->id)->first();
-                        $a->delete();
+                        $a =  Kpi::query()->where('task_id', $task->id)->where('stage_id', $stage->id)->first();
+                        if ($a != null) {
+                            $a->delete();
+                        }
                     }
 
                     HistoryMoveTask::query()->create([
@@ -125,12 +127,14 @@ class TaskController extends Controller
                     }else {
                         $j =0;
                     }
-                    Kpi::query()->create([
-                        'account_id' => $task->account_id,
-                        'stage_id' => $task->stage_id,
-                        'task_id' => $task->id,
-                        'status' => $j
-                    ]);
+                   if ($task->account_id != null) {
+                       Kpi::query()->create([
+                           'account_id' => $task->account_id,
+                           'stage_id' => $task->stage_id,
+                           'task_id' => $task->id,
+                           'status' => $j
+                       ]);
+                   }
                 }
 
                 $data['expired'] = null;
